Removed meaningless 'const' qualifiers
The Intel compiler generates a warning for 'meaningless' const
qualifiers. A const qualifier on an elemental return type, such
as:

const int foo();

carries no meaning because a value is being returned, not a pointer
or an object that could be changed. Three instances of this
were changed:

const size_t GridDimenensions::dimensionCount() (in repast_hpc/Point.h)
const Status AgentStatus::getStatus() (in repast_hpc/AgentStatus.h)
const int Observer::rank() (in relogo/Observer.h)

Note that the 'Status' being returned in the second example is an
enumerated type, which is effectively elemental.