Clearer Error Messages
The intent of this revision is to provide clearer error messages;
one motivation for this is to enable users to more easily diagnose
problems with their code through the output generated at runtime.

This extended into a larger framework, wherein errors are listed
in a header file (one for Repast, one for ReLogo), and constructed
using macros and some other syntactic sugar, so that the header
file contains something close to a human-readable reference on each
error. The output messages can be expanded to include values that
caused the errors (much more easily than in the original code),
notes about likely root causes for the errors, and steps to
avoid the errors. These can be updated as more feedback from
our users is gathered.