Refactor both Repast and ReLogo space hierarchy
Actually this is not a refactoring at all (as of this commit);
it is just a set of name changes.

Previously:

Class:                                    File:          Full Contents:
Projection                                Projection.h
   Grid                                   Grid.h
     BaseGrid                             BaseGrid.h     [GridPointHolder, BaseGrid]
       SharedBaseGrid                     SharedGrid.h   [CellContents, Neighbor, Neighbors, GridBufferSyncher, CartTopology, SharedBaseGrid]
          SharedGrid                      SharedSpace.h  [SharedGrid, SharedSpace, SharedGrids S:typedefs]
          SharedSpace                     SharedSpace.h

Plus one additional file:
Space.h S:typedefs

Proposal:

Class:                                    File:                     Full Contents:
Projection                                Projection.h
   Grid                                   Grid.h
     BaseGrid                             BaseGrid.h                [GridPointHolder, BaseGrid]
       S:typedefs                         Spaces.h
       SharedBaseGrid                     SharedBaseGrid.h          [CellContents, Neighbor, Neighbors, GridBufferSyncher, CartTopology, SharedBaseGrid]
          SharedDiscreteSpace             SharedDiscreteSpace.h     [SharedGridSpace]
          SharedContinuousSpace           SharedContinuousSpace.h   [SharedContinuousSpace]
               S: typedefs                SharedSpaces.h

The new hierarchy looks like this:

                           Projection
                               |
                              Grid
                               |
                            BaseGrid
                             /   \
                   (Spaces.h)     SharedBaseGrid
                                     /    \
                   SharedDiscreteSpace    SharedContinuousSpace
                                     \    /
                                 (SharedSpaces.h)

Note that the main changes are, first, that some confusing name changes are eliminated
(for example, the fact that the 'SharedGrid' class is not in SharedGrid.h). But the
main change is that the term 'Grid' is no longer used ambiguously. Previously it had
two meanings: a data structure from which spaces were made, and a kind of space. Now it
has just one meaning, the first of these. In place of the second, the term "DiscreteSpace"
is now used. What was previously called a 'Space' is now called a "ContinuousSpace".

This also required some changes to ReLogo, to the Spaces and Adders defined there.