BUG FIX
The method signature for:

AbstractRelogoAgent::turtlesOn(AgentSet<AgentType> out)

was modified to:

AbstractRelogoAgent::turtlesOn(AgentSet<AgentType>& out)

because a reference was required to collect the set of agents and
return it to the calling code. Note that this was an instrumental
part of the Zombies demo model that was presumably non-functional.