Modifications to allow manual specification of the order in which Projections are asked by a SharedContext to 'getAgentsToPush'.
The getAgentsToPush routine begins with a list of all local agents
and ends with a list of those agents that any projection thinks will
need to be 'pushed' to other projections.

This can be an inefficient operation. The data structures used by
some projections can be more or less amenable to finding agents
that must be 'pushed.'

However, once an agent is determined by any projection to be
a candidate for 'pushing', the other projections do not need
to test. The list of agents to test therefore reduces as each
projection marks sets of agents to be pushed.

Because of this, a strategic approach can be employed: have the
efficient projections do the tests first. They will take the 'easy'
agents out of the list, increasing (theoretically) the efficiency
of the overall routine.

Note that it is also possible that two projections can be 'matched':
they will always result in the same set of agents being 'pushed.'
This is common in the case of two correlated spatial projections. In
cases like this, one of the two projections- the inefficient one-
can be removed from the algorithm entirely.

The modifications here allow the specification of a vector of
projection names. The order can be manipulated and projections
removed if needed. The ReLogo implementation now automatically
removes the (inefficient) continuous spatial projection and relies
on the discrete space for 'pushing' all agents.

Future work will make this vector private and provide appropriate
getter/setter/manipulation methods.