Change to AgentImporterExporter.h and RepastProcess.h
Change pertains to condition in syncAgentStatuses; when sending
the exporter information needed so that receiving process knows
where to send copies of newly received local agents, the package
sent includes all export information, including exports that
were being sent to the receiving process itself. This was being
received by the receiving process and added to the ImporterExporter
(technically the Exporter), where it was appearing as a 'self'
export.

There is already a requirement that after syncAgentStatus was
called, syncProjectionInfo would be called, and
this immediately clears the ImporterExporter. However, in cases
where there are no projections, this holdover information
may cause problems. (I have not observed these, only suspect
them.)

To avert this I have added a way to delete the export information
held by the ImporterExporter for a specific destination process;
RepastProcess.h::syncAgentStatuses now calls this after moving
the exporter information into the IE.