Mark the source node in an edge as the 'master' node, if possible; if the source is nonlocal and the target is local, the target is marked as the 'master'. The edge will follow the 'master' node when the node is moved from one process to another. Note that creating an edge where both nodes are non-local will work (source will be marked as 'master'), but the node will not be moved to another process and will be deleted when projection information is synchronized.
The marking of source or target as 'master' happens automatically;
users cannot select which node is marked.

This change allows the creation of directed nodes with either
the source or target nonlocal. Previously the source node
was considered the 'master' node (this term was not used),
and creating a directed node from non-local source to local
target was effectively impossible because the edge would be deleted
on synchronization.