Prevents overwriting of duplicate edges, especially when two processes have (improperly!) created semantically duplicate edges that each believes to be a 'master' edge.
It is an error for two processes to create semantically duplicate
edges. Consider the case in which P0 borrows an agent from P1, and
P1 borrows an agent from P0. Both processes then create undirected
edges using the two agents involved. The source and target roles
may be reversed, but the edges are semantically the same because the
edge is undirected. When projection information is synchronized,
each process will attempt to send the edge to the process of its
non-local node, in the belief that the copy it has is the master.
The arriving edges will match the existing edge, and a collision
will occur.

Previously this collision was resolved in a counterintuitive and
less-than-useful way: the arriving edge always overwrote the
existing edge. The effect was that the two processes would
exchange their respective edges, though each would (still) believe
that it possessed the master edge.

This outcome has been replaced. Now if an incoming edge matches an
existing edge, the existing edge is kept and the arriving edge is
discarded. The existing edge is marked as 'conflicted'. All
RepastEdges now have a boolean value that indicates a conflicted
state. This value is NOT transmitted when the edge is sent across
processes. Graphs now have a function that clears this value for
all the edges they contain. When the synchronization routine is
run, extant edges that match incoming edges are marked as conflicte.
Graphs now have a function that returns the set of conflicted
edges. The intended use is: clear; synchronize; get list of conflicts.

The resolution of this conflict is up to the user. It should be
emphasized that this is a logical error in the simulation: two
processes should never both create semantically equal edges. However,
if a conflict arises, the user will be able to resolve it. In the
simple case of identical edges, a rule such as "delete the edge
with the higher process number for the local node" might suffice.
If the edges have additional properties, and these properties
differ between the two copies, the resolution may be more difficult.

Note also that conflicts can arise during synchronization of agent
statuses as well. When one process sends its agent to the other
process, the arriving agent comes with its link in tow, causing
another collision. The resolution in this case may be more difficult,
but it should be noted that the link on the process sending the agent
will persist as an invalid link (with two non-local nodes- the
original non-local node and the formerly local node, now changed
to non-local status). In theory the link could be recovered from
this.

Users are advised to avoid this problem by ensuring that their
simulations never create semantically duplicate edges.