A bug fix: During the synchronizeAgentStatuses routine a process (A) may receive notification that an agent that exists on that process (B) as a non-local agent is being moved from its home process to a new process. Two possibilities exist: the destination process may be a third process (C); or the destination process may be the process receiving the notification (A), meaning that the agent that has been a non-local agent will be received and will become a local agent.
Originally, the receiving process (A) would mark the agent with a new
ID only in the case of the first possibility- that the agent was going
to move to a third process (C). When this is the case, the agent must
be marked with the new ID, because this is the only time that the
agent ID will be updated. In the second case (the agent is moving
to process A) the routine did _not_ mark the agent with a new ID,
and instead relied on the fact that later in the process the agent
would be received in full (including its new ID) from the sending
process. Hence it was assumed that there was no need to mark the
agent when the notification was initially received.

However, testing revealed that this introduces a bug. If the
non-local agent on process (A) is a member of a master network link,
and the local agent that is the other node in that link is also
being moved to another process, then the agent will be sent as a
secondary agent along with the local agent to the process receiving
the local agent. In some cases, this will be process (B)- the
process where the agent was originally located as a local agent.

In this circumstance, the transmitted secondary agent has the old
ID value. This causes the receiving process to try to find the agent
as if it were a local agent. However, it has already been removed
because it is being moved to a new process. The result is that the
agent cannot be found; a segmentation fault is caused.

The solution in place here is to mark all agents with new IDs
when notification of movement is received. This will carry a
performance hit, but seems better than trying to determine if the
non-local agent may be a secondary agent.