Remove code from the original synchronization algorithm that has been deprecated and is no longer used in the new algorithm.
The primary motivation for this change is the confusion that the old code can cause among users who inspect the original
code closely. Although many of the routines were marked as being deprecated, they may have been more than obsolete: they
may actually have been harmful if run in conjunction with the new routines. Hence it is better to remove them entirely.

Among the classes that are being removed are such items as CellContents and GridBufferSyncher; essentially all of the functions
of these classes are now taken over by synchronizeProjectionInfo calls.