Correction to algorithm for transorming wraparound borders in double-precision spaces.
The previous algorithm would cause a crash if the double-precision coordinate was not
correctly dealt with in wraparound borders. In the canonical test case, an agent in a
space from [-100, 100) would proceed from -95 toward the lower bound by intervals of 1.2. In
double-precision c++ this would not arrive exactly at -100, but instead at some very small
distance past the boundary (e.g. -100.0000000000001). This is not in itself problematic,
but the transformation would be handled incorrectly; the agent should be placed at
99.9999999999999 but would instead be placed at exactly 100.000000, which is an
invalid value. This caused a segmentation fault when the agent was not found inside the
local boundaries but was also not within a buffer zone either.

The fix for this checks to see if the range is valid, and if not uses the c++
'nextafter' function to ensure that the value assigned to the agent is valid. Technically
there is a loss of precision, because an agent that should be placed at 99.9999999986
might end up at 99.999999999999 (or equivalent), but this is not preventable (because
adding such small precision values does not always lead to the correct result anyway)
and it is hoped that it is not significant.