Removed swapXY from CartTopology class.
swapXY was a function that took a vector of length 2 and exchanged
the first and second elements:

[ A, B] -> [B, A]

It was used within the CartTopology class to swap the dimensions
in the topology from the Repast HPC given dimensions to the MPI
Cartesian Topology dimensions. If the Repast HPC space were to be
16 x 8, the MPI Cartesian topology was converted to 8 x 16.

The reasons for this were never clear from the code, although there
was a comment about making it 'row major'. In fact, there were
several points where this swap was used, but the line between
Repast HPC's topology and the MPI Cartesian topology was
never crossed and hence never relevant: the functions simply
swapped everything going in or out, and hence were never really
needed.

The changes committed here remove the swap from the CartesianTopology
constructor and from the getDimensions method call, the two places
where it was called explicitly. Additionally, in the CreateNeighbors
method, the retrieval of the N/S and E/W adjacent ranks was changed from
the unusual E/W = dimension 1, N/S = dimension 0 back to the
expected E/W = dimension 0, N/S = dimension 1.